+++ src/app/pages/token/pages/new/introduction/introduction.component.html
@@ -37,7 +37,7 @@
                 i18n-nzTitle
                 nzTitle="Application form"
                 i18n-nzDescription
-                nzDescription="To apply you need to fill form below."
+                nzDescription="To apply you need to fill in the form below."
             >
             </nz-card-meta>
