+++ src/app/pages/token/pages/token/token.page.html
@@ -77,3 +77,4 @@
       </div>
     </nz-drawer>
   </wen-content>
+
