+++ src/app/pages/token/token.module.ts
@@ -14,6 +14,7 @@ import { TokenBidModule } from '@components/token/components/token-bid/token-bid
 import { TokenOfferPreMintModule } from '@components/token/components/token-offer-pre-mint/token-offer-pre-mint.module';
 import { TokenPublicSaleModule } from '@components/token/components/token-public-sale/token-public-sale.module';
 import { TokenPurchaseModule } from '@components/token/components/token-purchase/token-purchase.module';
+import { TokenRefundModule } from '@components/token/components/token-refund/token-refund.module';
 import { MarkDownModule } from '@core/pipes/markdown/markdown.module';
 import { TruncateModule } from '@core/pipes/truncate/truncate.module';
 import { LayoutModule } from '@shell/ui/layout/layout.module';
@@ -113,6 +114,7 @@ import { TokenRoutingModule } from './token-routing.module';
     TokenPublicSaleModule,
     NzSkeletonModule,
     TokenBidModule,
+    TokenRefundModule,
     TokenOfferPreMintModule,
     ModalDrawerModule,
     NgChartsModule
